extract position to hoc
