+++ apps/hmm1/src/app/Menu/CampaignMenu/CampaignMenu.tsx
@@ -1,5 +1,4 @@
-import { Button, type ButtonAssets, Menu, MenuItem } from '../../base';
-import type { PositionProps } from '../../core';
+import { Button, type ButtonAssets, Menu, MenuItem, type PositionProps } from '../../base';
 
 import cancelDisabled from './assets/cancel/disabled.png';
 import cancelEnabled from './assets/cancel/enabled.png';
