+++ apps/hmm1/src/app/Menu/ComPortMenu/ComPortMenu.tsx
@@ -1,7 +1,6 @@
 import { range } from 'lodash';
 
-import { Button, type ButtonAssets, Menu, MenuItem } from '../../base';
-import type { PositionProps } from '../../core';
+import { Button, type ButtonAssets, Menu, MenuItem, type PositionProps } from '../../base';
 
 import cancelDisabled from './assets/cancel/disabled.png';
 import cancelEnabled from './assets/cancel/enabled.png';
