--- apps/hmm1/src/app/Menu/ComPortMenu/assets/index.ts
@@ -1,13 +1,0 @@
-import { cancel } from "./cancel";
-import { com1 } from "./com-1";
-import { com2 } from "./com-2";
-import { com3 } from "./com-3";
-import { com4 } from "./com-4";
-
-export const buttonImages = {
-  cancel,
-  com1,
-  com2,
-  com3,
-  com4,
-};
