+++ apps/hmm1/src/app/Menu/MultiPlayerGameTypeMenu/MultiPlayerGameTypeMenu.tsx
@@ -1,5 +1,4 @@
-import { Button, Menu, MenuItem } from '../../base';
-import type { PositionProps } from '../../core';
+import { Button, Menu, MenuItem, type PositionProps } from '../../base';
 
 import cancelDisabled from './assets/cancel/disabled.png';
 import cancelEnabled from './assets/cancel/enabled.png';
