+++ apps/hmm1/src/app/base/Button/Button.tsx
@@ -1,15 +1,16 @@
 import { useState } from 'react';
 import styled from 'styled-components';
 
-import type { PositionProps } from '../../core';
+import { withPosition } from '../withPosition';
 
 export interface ButtonAssets {
   readonly disabled: string;
   readonly enabled: string;
 }
 
-interface Props extends PositionProps {
+interface Props {
   readonly assets: ButtonAssets;
+  readonly className?: string;
   readonly disabled?: boolean;
   readonly label: string;
   readonly onClick?: () => void;
@@ -17,60 +18,58 @@ interface Props extends PositionProps {
   readonly onMouseOver?: () => void;
 }
 
-export const Button = ({ assets, disabled, label, onClick, onMouseLeave, onMouseOver, x, y }: Props) => {
-  const [over, setOver] = useState(false);
-  const [pressed, setPressed] = useState(false);
+export const Button = withPosition(
+  ({ assets, className, disabled, label, onClick, onMouseLeave, onMouseOver }: Props) => {
+    const [over, setOver] = useState(false);
+    const [pressed, setPressed] = useState(false);
 
-  const handleMouseEnter = () => setOver(true);
+    const handleMouseEnter = () => setOver(true);
 
-  const handleMouseOut = () => setOver(false);
+    const handleMouseOut = () => setOver(false);
 
-  const handleDocumentMouseUp = () => {
-    document.removeEventListener('mouseup', handleDocumentMouseUp);
+    const handleDocumentMouseUp = () => {
+      document.removeEventListener('mouseup', handleDocumentMouseUp);
 
-    setPressed(false);
-  };
+      setPressed(false);
+    };
 
-  const handleMouseDown = () => {
-    document.addEventListener('mouseup', handleDocumentMouseUp);
+    const handleMouseDown = () => {
+      document.addEventListener('mouseup', handleDocumentMouseUp);
 
-    setPressed(true);
-  };
+      setPressed(true);
+    };
 
-  const handleMouseUp = () => {
-    document.removeEventListener('mouseup', handleDocumentMouseUp);
+    const handleMouseUp = () => {
+      document.removeEventListener('mouseup', handleDocumentMouseUp);
 
-    setPressed(false);
-  };
+      setPressed(false);
+    };
 
-  return (
-    <Root
-      aria-label={label}
-      disabled={disabled}
-      onClick={onClick}
-      onMouseDown={handleMouseDown}
-      onMouseEnter={handleMouseEnter}
-      onMouseLeave={onMouseLeave}
-      onMouseOut={handleMouseOut}
-      onMouseOver={onMouseOver}
-      onMouseUp={handleMouseUp}
-      x={x}
-      y={y}
-    >
-      <Image src={(over && pressed) || disabled ? assets.disabled : assets.enabled} />
-    </Root>
-  );
-};
+    return (
+      <Root
+        aria-label={label}
+        className={className}
+        disabled={disabled}
+        onClick={onClick}
+        onMouseDown={handleMouseDown}
+        onMouseEnter={handleMouseEnter}
+        onMouseLeave={onMouseLeave}
+        onMouseOut={handleMouseOut}
+        onMouseOver={onMouseOver}
+        onMouseUp={handleMouseUp}
+      >
+        <Image src={(over && pressed) || disabled ? assets.disabled : assets.enabled} />
+      </Root>
+    );
+  },
+);
 
-const Root = styled.button<Pick<Props, 'x' | 'y'>>(({ x, y }) => ({
+const Root = styled.button({
   backgroundColor: 'transparent',
   border: 0,
   fontSize: 0,
-  left: x,
   padding: 0,
-  position: 'absolute',
-  top: y,
-}));
+});
 
 const Image = styled.img({
   userDrag: 'none',
