+++ apps/hmm1/src/app/core.ts
@@ -1,7 +1,2 @@
 export const ScreenWidth = 640;
 export const ScreenHeight = 480;
-
-export interface PositionProps {
-  readonly x?: number;
-  readonly y?: number;
-}
